CLN: Move Clickhouse Client class to Backend
