+++ ibis/backends/clickhouse/client.py
@@ -1,19 +1,11 @@
 import re
-from collections import OrderedDict
 
 import numpy as np
 import pandas as pd
-from clickhouse_driver.client import Client as _DriverClient
 
-import ibis
 import ibis.common.exceptions as com
 import ibis.expr.datatypes as dt
-import ibis.expr.schema as sch
 import ibis.expr.types as ir
-from ibis.backends.base.sql import SQLClient
-from ibis.config import options
-
-from .compiler import ClickhouseCompiler
 
 fully_qualified_re = re.compile(r"(.*)\.(?:`(.*)`|(.*))")
 base_typename_re = re.compile(r"(\w+)")
@@ -157,115 +149,3 @@ class ClickhouseTable(ir.TableExpr):
 
         data = obj.to_dict('records')
         return self._client.con.execute(query, data, **kwargs)
-
-
-class ClickhouseClient(SQLClient):
-    """An Ibis client interface that uses Clickhouse"""
-
-    compiler = ClickhouseCompiler
-
-    def __init__(self, backend, *args, **kwargs):
-        self.backend = backend
-        self.database_class = backend.database_class
-        self.table_class = backend.table_class
-        self.table_expr_class = backend.table_expr_class
-        self.con = _DriverClient(*args, **kwargs)
-
-    def raw_sql(self, query: str, external_tables={}):
-        external_tables_list = []
-        for name, df in external_tables.items():
-            if not isinstance(df, pd.DataFrame):
-                raise TypeError(
-                    'External table is not an instance of pandas ' 'dataframe'
-                )
-            schema = sch.infer(df)
-            external_tables_list.append(
-                {
-                    'name': name,
-                    'data': df.to_dict('records'),
-                    'structure': list(
-                        zip(
-                            schema.names,
-                            [
-                                str(ClickhouseDataType.from_ibis(t))
-                                for t in schema.types
-                            ],
-                        )
-                    ),
-                }
-            )
-
-        ibis.util.log(query)
-        return self.con.execute(
-            query,
-            columnar=True,
-            with_column_types=True,
-            external_tables=external_tables_list,
-        )
-
-    def ast_schema(self, query_ast, external_tables={}):
-        # Allowing signature to accept `external_tables`
-        return super().ast_schema(query_ast)
-
-    def fetch_from_cursor(self, cursor, schema):
-        data, columns = cursor
-        if not len(data):
-            # handle empty resultset
-            return pd.DataFrame([], columns=schema.names)
-
-        df = pd.DataFrame.from_dict(OrderedDict(zip(schema.names, data)))
-        return schema.apply_to(df)
-
-    def close(self):
-        """Close Clickhouse connection and drop any temporary objects"""
-        self.con.disconnect()
-
-    def _fully_qualified_name(self, name, database):
-        if bool(fully_qualified_re.search(name)):
-            return name
-
-        database = database or self.current_database
-        return f'{database}.`{name}`'
-
-    def get_schema(self, table_name, database=None):
-        """
-        Return a Schema object for the indicated table and database
-
-        Parameters
-        ----------
-        table_name : string
-          May be fully qualified
-        database : string, default None
-
-        Returns
-        -------
-        schema : ibis Schema
-        """
-        qualified_name = self._fully_qualified_name(table_name, database)
-        query = f'DESC {qualified_name}'
-        data, columns = self.raw_sql(query)
-        return sch.schema(
-            data[0], list(map(ClickhouseDataType.parse, data[1]))
-        )
-
-    def set_options(self, options):
-        self.con.set_options(options)
-
-    def reset_options(self):
-        # Must nuke all cursors
-        raise NotImplementedError
-
-    def _ensure_temp_db_exists(self):
-        name = (options.clickhouse.temp_db,)
-        if name not in self.list_databases():
-            self.create_database(name, force=True)
-
-    def _get_schema_using_query(self, query, **kwargs):
-        data, columns = self.raw_sql(query, **kwargs)
-        colnames, typenames = zip(*columns)
-        coltypes = list(map(ClickhouseDataType.parse, typenames))
-        return sch.schema(colnames, coltypes)
-
-    def _table_command(self, cmd, name, database=None):
-        qualified_name = self._fully_qualified_name(name, database)
-        return f'{cmd} {qualified_name}'
