+++ ibis/backends/postgres/__init__.py
@@ -1,10 +1,11 @@
 """PostgreSQL backend."""
+from ibis.backends.base import BaseBackend
 from ibis.backends.base.sql.alchemy import BaseAlchemyBackend
 
 from .client import PostgreSQLClient
 
 
-class Backend(BaseAlchemyBackend):
+class Backend(BaseAlchemyBackend, BaseBackend):
     name = 'postgres'
     client_class = PostgreSQLClient
